pyproject.toml
[tool.poetry]
name = "etlprojectapiextract"
version = "0.1.0"
description = ""
authors = ["Raul.SantosSilveira-ext <Raul.SantosSilveira-ext@ab-Inbev.com >"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
tinydb = "^4.8.2"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
logfire = {extras = ["requests", "sqlalchemy"], version = "^2.9.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
